Day 23 (November 30, 2021)
November 30, 2021 is is the day I first started work on p5js-tetrominoes.
The first semester of my Intro to Engineering class was a coding class.
This initial commit represents the progress I made on the very first
in-class session of working on what was going to be my final project for
the semester.

If you're wondering why the day-numbering scheme is weird, I'll explain.
Day 1 was our very first session of instruction with the coding instructor,
and the day counter incremented every time we were with the instructor.
When the day counter got an A on the end it meant that we were still
working on p5.js but we were not with the instructor. Those "A" days were
usually Fridays. If you're wondering why there was an "A" day every 2
normal days instead of 4, our school has a block schedule where we have
half our classes on half of the days as a sacrifice for having longer
periods, with the exception of Fridays.

You may also notice that the name of the game I'm trying to clone has been
censored due to fear of legal action with the T****s Company.